pttbackend/requirements.txt
django==2.2.10
graphene-django==2.8.1
django-cors-headers==3.2.1
django-graphql-jwt==0.3.0
